refactor: split stats/db layers and unify AnalysisOptions in reddwarf-ts
- Pure statistical functions stay in stats.ts; DB types (VoteConnection,
  VoteQueryResult) and getGroupVoteMatrices move to new db.ts
- Collapse analyzeLabeledGroups + calculateRepresentativeStatements into
  one function in representative-statements.ts; stats.ts is now DB-free
- Unified AnalysisOptions covers all options incl. commentTextMap;
  removes two divergent options shapes across the package
- App adapter updated: commentTextMap folds into options object
