+++ packages/reddwarf-ts/src/index.ts
@@ -1,3 +1,4 @@
 export * from './stats.js';
+export * from './db.js';
 export * from './representative-statements.js';
 export * from './druid-reducer.js';
